refactor utils -> examples
