+++ bxsolana/examples/__init__.py
@@ -4,3 +4,13 @@ from .stream_utils import do_stream
 from .constants import PUBLIC_KEY, USDC_WALLET, OPEN_ORDERS, ORDER_ID, SOL_USDC_MARKET
 from .order_utils import cancel_order, replace_order_by_client_order_i_d
 from .order_lifecycle import order_lifecycle
+
+__all__ = [
+    "do_requests",
+    "do_transaction_requests",
+    "do_stream",
+    "cancel_order",
+    "replace_order_by_client_order_i_d",
+    "order_lifecycle",
+    "PUBLIC_KEY", "USDC_WALLET", "OPEN_ORDERS", "ORDER_ID", "SOL_USDC_MARKET",
+]
