+++ example/transaction/main.py
@@ -3,7 +3,7 @@ import os
 
 from bxsolana_trader_proto import api as proto
 from bxsolana import provider
-from bxsolana import utils
+from bxsolana import examples
 
 # TODO: Add some logic here to indicate to user if missing needed environment variables for tests
 public_key = os.getenv("PUBLIC_KEY")
@@ -29,97 +29,97 @@ async def ws():
     print("\n*** WS Test ***\n")
     async with provider.ws() as api:
         async with provider.ws() as api2:  # TODO use same provider when WS streams are separated
-            await utils.order_lifecycle(p1=api,
-                                        p2= api2,
-                                        owner_addr=public_key,
-                                        payer_addr=public_key,
-                                        market_addr=market_addr,
-                                        order_side=order_side,
-                                        order_type=order_type,
-                                        order_amount=order_amount,
-                                        order_price=order_price,
-                                        open_orders_addr=open_orders,
-                                        base_token_wallet=base_token_wallet,
-                                        quote_token_wallet=quote_token_wallet)
+            await examples.order_lifecycle(p1=api,
+                                           p2= api2,
+                                           owner_addr=public_key,
+                                           payer_addr=public_key,
+                                           market_addr=market_addr,
+                                           order_side=order_side,
+                                           order_type=order_type,
+                                           order_amount=order_amount,
+                                           order_price=order_price,
+                                           open_orders_addr=open_orders,
+                                           base_token_wallet=base_token_wallet,
+                                           quote_token_wallet=quote_token_wallet)
 
-            await utils.cancel_all_orders(api,
-                                          owner_addr=public_key,
-                                          payer_addr=public_key,
-                                          order_side=order_side,
-                                          order_type=order_type,
-                                          order_amount=order_amount,
-                                          order_price=order_price,
-                                          open_orders_addr=open_orders,
-                                          market_addr=market_addr)
+            await examples.cancel_all_orders(api,
+                                             owner_addr=public_key,
+                                             payer_addr=public_key,
+                                             order_side=order_side,
+                                             order_type=order_type,
+                                             order_amount=order_amount,
+                                             order_price=order_price,
+                                             open_orders_addr=open_orders,
+                                             market_addr=market_addr)
 
-            await utils.replace_order_by_client_order_i_d(api,
-                                                          owner_addr=public_key,
-                                                          payer_addr=public_key,
-                                                          market_addr=market_addr,
-                                                          order_side=order_side,
-                                                          order_type=order_type,
-                                                          order_amount=order_amount,
-                                                          order_price=order_price,
-                                                          open_orders_addr=open_orders)
+            await examples.replace_order_by_client_order_i_d(api,
+                                                             owner_addr=public_key,
+                                                             payer_addr=public_key,
+                                                             market_addr=market_addr,
+                                                             order_side=order_side,
+                                                             order_type=order_type,
+                                                             order_amount=order_amount,
+                                                             order_price=order_price,
+                                                             open_orders_addr=open_orders)
 
 async def grpc():
     print("\n*** GRPC Test ***\n")
     async with provider.grpc() as api:
-        await utils.order_lifecycle(p1=api,
-                                    p2=api,
-                                    owner_addr=public_key,
-                                    payer_addr=public_key,
-                                    market_addr=market_addr,
-                                    order_side=order_side,
-                                    order_type=order_type,
-                                    order_amount=order_amount,
-                                    order_price=order_price,
-                                    open_orders_addr=open_orders,
-                                    base_token_wallet=base_token_wallet,
-                                    quote_token_wallet=quote_token_wallet)
+        await examples.order_lifecycle(p1=api,
+                                       p2=api,
+                                       owner_addr=public_key,
+                                       payer_addr=public_key,
+                                       market_addr=market_addr,
+                                       order_side=order_side,
+                                       order_type=order_type,
+                                       order_amount=order_amount,
+                                       order_price=order_price,
+                                       open_orders_addr=open_orders,
+                                       base_token_wallet=base_token_wallet,
+                                       quote_token_wallet=quote_token_wallet)
 
-        await utils.cancel_all_orders(api,
-                                      owner_addr=public_key,
-                                      payer_addr=public_key,
-                                      order_side=order_side,
-                                      order_type=order_type,
-                                      order_amount=order_amount,
-                                      order_price=order_price,
-                                      open_orders_addr=open_orders,
-                                      market_addr=market_addr)
+        await examples.cancel_all_orders(api,
+                                         owner_addr=public_key,
+                                         payer_addr=public_key,
+                                         order_side=order_side,
+                                         order_type=order_type,
+                                         order_amount=order_amount,
+                                         order_price=order_price,
+                                         open_orders_addr=open_orders,
+                                         market_addr=market_addr)
 
-        await utils.replace_order_by_client_order_i_d(api,
-                                                      owner_addr=public_key,
-                                                      payer_addr=public_key,
-                                                      market_addr=market_addr,
-                                                      order_side=order_side,
-                                                      order_type=order_type,
-                                                      order_amount=order_amount,
-                                                      order_price=order_price,
-                                                      open_orders_addr=open_orders)
+        await examples.replace_order_by_client_order_i_d(api,
+                                                         owner_addr=public_key,
+                                                         payer_addr=public_key,
+                                                         market_addr=market_addr,
+                                                         order_side=order_side,
+                                                         order_type=order_type,
+                                                         order_amount=order_amount,
+                                                         order_price=order_price,
+                                                         open_orders_addr=open_orders)
 
 async def http():
     print("\n*** HTTP Test ***\n")
     async with provider.http() as api:
-        await utils.cancel_all_orders(api,
-                                      owner_addr=public_key,
-                                      payer_addr=public_key,
-                                      order_side=order_side,
-                                      order_type=order_type,
-                                      order_amount=order_amount,
-                                      order_price=order_price,
-                                      open_orders_addr=open_orders,
-                                      market_addr=market_addr)
+        await examples.cancel_all_orders(api,
+                                         owner_addr=public_key,
+                                         payer_addr=public_key,
+                                         order_side=order_side,
+                                         order_type=order_type,
+                                         order_amount=order_amount,
+                                         order_price=order_price,
+                                         open_orders_addr=open_orders,
+                                         market_addr=market_addr)
 
-        await utils.replace_order_by_client_order_i_d(api,
-                                                      owner_addr=public_key,
-                                                      payer_addr=public_key,
-                                                      market_addr=market_addr,
-                                                      order_side=order_side,
-                                                      order_type=order_type,
-                                                      order_amount=order_amount,
-                                                      order_price=order_price,
-                                                      open_orders_addr=open_orders)
+        await examples.replace_order_by_client_order_i_d(api,
+                                                         owner_addr=public_key,
+                                                         payer_addr=public_key,
+                                                         market_addr=market_addr,
+                                                         order_side=order_side,
+                                                         order_type=order_type,
+                                                         order_amount=order_amount,
+                                                         order_price=order_price,
+                                                         open_orders_addr=open_orders)
 
 if __name__ == "__main__":
     asyncio.run(main())
